+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add tests to ensure inline plugin works <https://github.com/gtronset/beets-filetote/pull/130>
+- Bugfix for Update CLI command <https://github.com/gtronset/beets-filetote/pull/133>
 
 ## [0.4.6] - 2023-12-03